decouple code and css from template
