+++ src/components/nav/nav.js
@@ -1,0 +1,12 @@
+
+export default {
+  data() {
+    return {}
+  },
+  props: {
+    commentNum: {
+      type: Number,
+      default: 0
+    }
+  }
+};
