+++ src/components/order/order.vue
@@ -12,34 +12,5 @@
 </van-list>
 </template>
 
-<script>
-import axios from 'axios';
-export default {
-  data() {
-    return {
-      list: [],
-      loading: false,
-      finished: false,
-    };
-  },
-  methods: {
-    onLoad() {
-      var that = this;
-          axios.post('http://localhost:6166/payment/page',{
-          
-        })
-        .then((res) => {
-              if (res.data.code === 200) {
-                that.list = res.data.data;
-                that.finished = true;
-              } else {
-              }
-            })
-    },
-  },
-};
-</script>
-
-<style>
-
-</style>
+<script src="./order.js"></script>
+<style src="./order.css></style>
