+++ src/components/split/Split.vue
@@ -2,14 +2,5 @@
   <div class="split"></div>
 </template>
 
-<script>
-  export default {}
-</script>
-
-<style>
-  .split{
-    width: 100%;
-    height: 10px;
-    background: #F4F4F4;
-  }
-</style>
+<script src="./split.js"></script>
+<style  src="./split.css></style>
