+++ src/components/split/split.css
@@ -1,0 +1,5 @@
+.split{
+  width: 100%;
+  height: 10px;
+  background: #F4F4F4;
+}
