+++ src/components/split/split.js
@@ -1,0 +1,2 @@
+
+export default {}
