+++ src/components/star/star.css
@@ -1,0 +1,25 @@
+.star{
+  font-size: 0;
+}
+.star .star-item{
+  display: inline-block;
+  width: 10px;
+  height: 10px;
+  margin-right: 3px;
+  background-repeat: no-repeat;
+  background-size: 10px 10px;
+}
+.star .star-item:last-child{
+  margin-right: 0;
+}
+
+/* 三种图片类型*/
+.star .on{
+  background-image: url(img/star24_on@2x.png);
+}
+.star .half{
+  background-image: url(img/star24_half@2x.png);
+}
+.star .off{
+  background-image: url(img/star24_off@2x.png);
+}
